requirements.txt
asgiref==3.4.0
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
djangorestframework==3.12.4
html5lib==1.1
idna==2.10
pycodestyle==2.7.0
pymemcache==3.4.4
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
webencodings==0.5.1
